Shorten the function length
